refactor(client): file-structure
