+++ client/src/views/components/editor/typeSettings/index.ts
