+++ client/src/views/components/editor/typeSettings/typeOptionProps.ts
