+++ client/src/views/elements/editor/typeEditors/color.tsx
@@ -1,5 +1,5 @@
 import * as React                 from 'react';
-import { ColorTextInput }         from '../../../../../util/index';
+import { ColorTextInput }         from '../../../../util/index';
 import { TypeElementEditorProps } from './typeEditorProps';
 
 export default ({ field, record, onDataChange }: TypeElementEditorProps<'color'>) => (
