+++ client/src/views/elements/editor/typeEditors/tags.tsx
@@ -1,5 +1,5 @@
 import * as React                 from 'react';
-import { TagInput }               from '../../../../../util/index';
+import { TagInput }               from '../../../../util/index';
 import { TypeElementEditorProps } from './typeEditorProps';
 
 export default ({ field, record, onDataChange }: TypeElementEditorProps<'tags'>) => (
