+++ client/src/views/elements/editor/typeEditors/text.tsx
@@ -1,13 +1,12 @@
 import * as React                 from 'react';
-import { SimpleTextField }        from '../../../../../util/index';
+import { SimpleTextField }        from '../../../../util/index';
 import { TypeElementEditorProps } from './typeEditorProps';
 
-export default ({ field, record, onDataChange }: TypeElementEditorProps<'textArea'>) => (
+export default ({ field, record, onDataChange }: TypeElementEditorProps<'text'>) => (
     <SimpleTextField
         TextFieldProps={{
             fullWidth: true
         }}
-        multiline={true}
         label={field.label}
         value={record as string}
         onBlur={onDataChange}
