+++ client/src/views/elements/editor/typeEditors/textArea.tsx
@@ -1,12 +1,13 @@
 import * as React                 from 'react';
-import { SimpleTextField }        from '../../../../../util/index';
+import { SimpleTextField }        from '../../../../util/index';
 import { TypeElementEditorProps } from './typeEditorProps';
 
-export default ({ field, record, onDataChange }: TypeElementEditorProps<'text'>) => (
+export default ({ field, record, onDataChange }: TypeElementEditorProps<'textArea'>) => (
     <SimpleTextField
         TextFieldProps={{
             fullWidth: true
         }}
+        multiline={true}
         label={field.label}
         value={record as string}
         onBlur={onDataChange}
